+++ src/fontrenderer.h
@@ -51,10 +51,10 @@ public:
     ~FontRenderer();
 
     const QVector<LayoutChar>& rendered() const { return m_chars;}
-    void placeImage(QPainter& p,ushort sybol,int x,int y);
+    void placeImage(QPainter& p,uint sybol,int x,int y);
     const RendererData& data() const { return m_rendered;}
     void LockAll();
-    void SetImage(ushort symb,const QImage& img);
+    void SetImage(uint symb,const QImage& img);
 private:
     const FontConfig* m_config;
     FT_Library m_ft_library;
@@ -64,8 +64,8 @@ private:
     RendererData m_rendered;
     QVector<LayoutChar> m_chars;
     void clear_bitmaps();
-    bool append_bitmap(ushort symbol);
-    void append_kerning(ushort symbol,const ushort* other,int amount);
+    bool append_bitmap(uint symbol);
+    void append_kerning(uint symbol,const uint* other,int amount);
 signals:
     void imagesChanged();
     void imagesChanged(const QVector<LayoutChar>&);
